pyproject.toml
[tool.poetry]
name = "waka-box-py"
version = "0.1.0"
description = "将你的 WakaTime 每周语言使用统计显示在 profile README / pined gist"
authors = ["Karako <karakohear@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.1"
ujson = "^5.4.0"
pydantic = "^1.10.2"
alive-progress = "^2.4.1"
"github3.py" = "^3.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
